Changed range mode.

+++ api/rpApplications/lcr_meter/src/utils.h
@@ -34,7 +34,7 @@
 
 float vectorMax(float *data, int size);
 float vectorApprox(float *data, int size, float approx_val, bool min);
-float trapezoidalApprox(float *data, float t, int size);
+float trapezoidalApprox(double *data, float t, int size);
 float vectorMean(float *data, int steps);
 float **multiDimensionVector(int second_dimenson);
 int set_IIC_Shunt(int k);
